// src/app/modules/statistics/enum/projectStatus.enum.ts
export enum ProjectStatus {
  APROBADO = 'APROBADO',
  APROBADO_OBSERVACIONES = 'APROBADO_OBSERVACIONES',
  NO_APROBADO = 'NO_APROBADO',
  EN_REVISION = 'EN_REVISION',
  EN_DESARROLLO = 'EN_DESARROLLO', // Nuevo
  APLAZADO = 'APLAZADO',           // Nuevo
  SUSPENDIDO = 'SUSPENDIDO',       // Nuevo
  CANCELADO = 'CANCELADO'          // Nuevo
}